Fix wrong struct assignment usage in options_settable_test
Summary:

As the comment in options_settable_test.cc says, it based on the behavior of
compiler that padding bytes are not changed when copying the struct. Current
ColumnFamilyOptionsAllFieldsSettable test case uses struct assignment, which
is dangerous if struct layout changes in the future. See the following
demonstration code for an example:

/* demo.cpp */

#include <cstdio>
#include <cstring>
#include <memory>

static const char kSpecialChar = 'z';

struct SmallStruct {
  unsigned a;
  bool b;
  unsigned c;
  SmallStruct(): a(0), b(false), c(0) {}
};

struct BigStruct {
  SmallStruct a;
  SmallStruct b;
};

size_t CountSpecialChar(char* buf, size_t size) {
  size_t result = 0;
  for (size_t i = 0; i < size; i++) {
    result += *(buf + i) == kSpecialChar ? 1 : 0;
  }
  return result;
}

int main()
{
  char* buf = new char[sizeof(BigStruct)];

  memset(buf, kSpecialChar, sizeof(BigStruct));
  ((BigStruct*)(buf))->a = SmallStruct();
  ((BigStruct*)(buf))->b = SmallStruct();
  printf("%lu ", CountSpecialChar(buf, sizeof(BigStruct)));

  memset(buf, kSpecialChar, sizeof(BigStruct));
  new (buf) BigStruct();
  printf("%lu\n", CountSpecialChar(buf, sizeof(BigStruct)));

  return 0;
}

The demo code, compiled by g++-5.4.0, g++-7.4.0 or g++-8.1.0 on an x64 Linux
machine, will output "0 6", which shows that struct assignment is not safe in
the special character counting mechanism.

This commit changes struct assignment to new operations to avoid the issue.